Do not align stored blocks to a byte boundary.
This diverts further from the deflate standard, but it was not working
in the first place because the decoder has no concept of a byte boundary
when reading from a 47-bit chunk stream.

We also get rid of the stream checksum (16 bits), it is a bit overkill
for our use, and it helps us reduce the decoder size.

Code size: 2498
Compressed code size: 1238